Refactorizado el codigo para que funcione con la app de customers
