+++ customers/urls.py
@@ -1,0 +1,12 @@
+from django.conf.urls import url
+
+from . import views
+
+app_name = 'customers'
+
+urlpatterns = [
+    # Customer orders
+    url(r'^customers/orders/$', views.customer_orders, name='new_customer_order'),
+    url(r'^customers/orders/new/$', views.new_customer_order, name='new_customer_order'),
+
+]
